Refactor FlatsCommand into modular subcommands
Extracted functionalities from FlatsCommand into separate SubCommand classes for modularity and maintainability. Updated the project structure to reflect subcommands as independent entities. Introduced utility classes for shared operations to reduce code duplication.
