+++ build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 }
 
 group = "de.nvclas"
-version = "0.6.1"
+version = "0.6.5"
 
 repositories {
     mavenCentral()
