+++ src/main/java/de/nvclas/flats/commands/flats/subcommands/UpdateSubCommand.java
@@ -1,0 +1,37 @@
+package de.nvclas.flats.commands.flats.subcommands;
+
+import de.nvclas.flats.Flats;
+import de.nvclas.flats.commands.flats.SubCommand;
+import de.nvclas.flats.updater.UpdateDownloader;
+import de.nvclas.flats.updater.UpdateStatus;
+import de.nvclas.flats.util.I18n;
+import de.nvclas.flats.util.Permissions;
+import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
+
+public class UpdateSubCommand implements SubCommand {
+    
+    private final Flats flatsPlugin;
+    
+    public UpdateSubCommand(Flats flatsPlugin) {
+        this.flatsPlugin = flatsPlugin;
+    }
+    
+    @Override
+    public void execute(@NotNull Player player, @NotNull String @NotNull [] args) {
+        if (Permissions.hasNoPermission(player, Permissions.ADMIN)) {
+            return;
+        }
+        UpdateDownloader updateDownloader = new UpdateDownloader(flatsPlugin,
+                "https://api.github.com/repos/nvclas/Flats/releases/latest");
+        UpdateStatus status = updateDownloader.downloadLatestRelease();
+        switch (status) {
+            case SUCCESS -> {
+                updateDownloader.unloadPluginAndDeleteJar();
+                player.sendMessage(Flats.PREFIX + I18n.translate("update.success", updateDownloader.getFileName()));
+            }
+            case NOT_FOUND -> player.sendMessage(Flats.PREFIX + I18n.translate("update.not_found"));
+            case FAILED -> player.sendMessage(Flats.PREFIX + I18n.translate("update.failed"));
+        }
+    }
+}
